MAINT: move remaining MaskedArray matrix tests to matrixlib.
